pyproject.toml
[tool.poetry]
name = "task-project"
version = "0.1.0"
description = ""
authors = ["Sayed1-11 <sheikhsayed0202@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
pytest = "^8.1.1"
pydantic = "^2.6.4"
django-ninja = "^1.1.0"
django-extensions = "^3.2.3"
django-autoslug = "^1.9.9"
pillow = "^10.2.0"
dj-database-url = "^2.1.0"
environ = "^1.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-celery-beat = "^2.6.0"
celery = "^5.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
